requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
chromadb>=0.4.15
groq>=0.4.0
python-dotenv>=1.0.0
pypdf>=3.17.0
tiktoken>=0.5.1
sentence-transformers>=2.2.0
torch>=2.0.0

